Removing userconfig from dxt manifest json for now
If the user doesn't specify kubeconfig in that configuration setting, the server just errors out if loaded via DXT. This fixes that so that it will use the default ~/.kube/config for now.

Another issue that I found is that setting KUBECONFIG=~/.kube/config doesn't actually expand "~" into home dir, so this fails with Claude Desktop.


Summary:

Test Plan: